Get donor fields in correct order
